rename crates
cosmogony_model -> cosmogony

root (cosmogony before) -> cosmogony_builder

and thus we'll still only publish cosmogony to crates.io
