+++ cosmogony/src/lib.rs
@@ -1,6 +1,6 @@
-mod cosmogony;
+mod model;
 pub mod mutable_slice;
 mod zone;
 
 pub use crate::zone::{Coord, Zone, ZoneIndex, ZoneType};
-pub use cosmogony::{Cosmogony, CosmogonyMetadata, CosmogonyStats};
+pub use model::{Cosmogony, CosmogonyMetadata, CosmogonyStats};
